feat: add option to gently create bin links/shims
This adds the top level `binLink(from, to, opts, cb)` method.

If `opts.gently` is a string, and `opts.clobberLinkGently` is set to
true, and `opts.force` is not set to true, then it will only create the
link or cmd shim if the current target either does not exist, or is a
link/shim into the path specified by `opts.gently`.

This will be used by the `bin-links` package to prevent top-level global
packages from overwriting bins that are not related to the package being
installed.

Paired with @mikemimik